requirements.txt
# ── Core (matches original) ────────────────────────────────────────────────────
streamlit>=1.35.0
boto3>=1.34.0
botocore>=1.34.0
pandas>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0

# ── AWS Strands Agents SDK (new in v2) ─────────────────────────────────────────
strands-agents>=1.0.0
strands-agents-tools>=0.1.0

# ── MCP Server support (new in v2) ────────────────────────────────────────────
mcp>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
pyarrow>=14.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
